Refactored map.js.
